Fixed 64bit integers on 32bit machines to resolve #527.
Free Software Free Society

To support commits by ctubio,
you can buy-me-a-drink with a small git tip at:
  1GitTipgxvKB3zjCLXRcSgDhC9pivkpc7u

I promise to drink chocolate milk in the development of the next commit.

To request new features or in case this commit breaks something for you,
please create a new github issue with all possible details,
but never share your API Keys!

+++ src/server/db.h
@@ -48,7 +48,7 @@ namespace K {
         sqlite3_free(zErrMsg);
         return j;
       };
-      void insert(mMatter k, json o, bool rm = true, string id = "NULL", unsigned long expire = 0) {
+      void insert(mMatter k, json o, bool rm = true, string id = "NULL", unsigned long long expire = 0) {
         ((EV*)events)->deferred([this, k, o, rm, id, expire]() {
           char* zErrMsg = 0;
           sqlite3_exec(db, (
