+++ src/server/pg.h
@@ -8,8 +8,8 @@ namespace K {
       map<double, mTrade> buys;
       map<double, mTrade> sells;
       map<string, mWallet> balance;
-      unsigned long profitT_21s = 0;
-      unsigned long walletT_2s = 0;
+      unsigned long long profitT_21s = 0;
+      unsigned long long walletT_2s = 0;
       string sideAPR_ = "!=";
     public:
       mPosition position;
@@ -197,7 +197,7 @@ namespace K {
         skip();
       };
       void expire(map<double, mTrade> *k) {
-        unsigned long now = _Tstamp_;
+        unsigned long long now = _Tstamp_;
         for (map<double, mTrade>::iterator it = k->begin(); it != k->end();)
           if (it->second.time + qp->tradeRateSeconds * 1e+3 > now) ++it;
           else it = k->erase(it);
@@ -294,7 +294,7 @@ namespace K {
         }
       }
       void calcProfit(mPosition *k) {
-        unsigned long now = _Tstamp_;
+        unsigned long long now = _Tstamp_;
         if (profitT_21s<=3) ++profitT_21s;
         else if (k->baseValue and k->quoteValue and profitT_21s+21e+3 < now) {
           profitT_21s = now;
